requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# Caching
redis==5.0.1
fastapi-cache2==0.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1
pytest==8.0.0
httpx==0.26.0
pytest-asyncio==0.23.5 